the macros
please forgive the use of for loop macros and . . . that other thing. They are not meant to stay in the code I was goofing around but has actually been to working too conveniently in testing to justify trashing them just yet. I'm really just trying to learn to collaboratively code with others so please submit any and all edit or critique.